requirements.txt
imageio==2.28.1
matplotlib==3.7.1
numpy==1.24.3
open3d==0.13.0
Pillow==9.5.0
Pillow==9.5.0
pycolmap==0.4.0
pyviz3d==0.2.32
scipy==1.10.1
torch==2.0.1
tqdm==4.65.0
transformers==4.29.1
